Fix hid_send_feature_report failure on 64-bit systems with 113a devices
The ctypes argtype for the size parameter of hid_send_feature_report was
declared as c_int instead of c_size_t, causing the 524-byte packet length
to be misinterpreted on 64-bit systems. This made the tool fail with
HIDSendError on MSI laptops using the newer SteelSeries KLC controller
(USB product ID 1038:113a), such as the GS76 Stealth.

Also updates the udev rule to match the 113a product ID.

+++ msi_perkeyrgb/hidapi_types.py
@@ -25,7 +25,7 @@ def set_hidapi_types(hidapi):
     hidapi.hid_read.restype = ct.c_int
     hidapi.hid_set_nonblocking.argtypes = [ct.c_void_p, ct.c_int]
     hidapi.hid_set_nonblocking.restype = ct.c_int
-    hidapi.hid_send_feature_report.argtypes = [ct.c_void_p, ct.c_char_p, ct.c_int]
+    hidapi.hid_send_feature_report.argtypes = [ct.c_void_p, ct.c_char_p, ct.c_size_t]
     hidapi.hid_send_feature_report.restype = ct.c_int
     hidapi.hid_get_feature_report.argtypes = [ct.c_void_p, ct.c_char_p, ct.c_size_t]
     hidapi.hid_get_feature_report.restype = ct.c_int
